Escape dots in service names
RabbitMQ does not support dots in topic names. Fully qualified service
names usually have dot characters in it and need to be escaped before
using as a topic name in the client and unescaped in the remote client.

I used hyphen as a replacement because hyphens are not valid
identifier characters according to protobuf specification.