Remove ProtoBuf encoding/decoding from RemoteClient (#39)
`RemoteClient` decodes protobuf messages received from the gRPC service to a Haskell type to immediately encode it back to protobuf and send it to MQTT. This consumes a significant amount of memory and CPU.

This change removes the encoding/decoding logic completely from the `RemoteClient`. Instead, we pass the protobuf encoded bytes directly between the `Client` and the remote gRPC service. Any serialization errors will be handled by them instead of the `RemoteClient`.